Moving the resource routers to a separate directory
