+++ src/routers/resource_routers/case_study_router.py
