+++ src/routers/resource_routers/dataset_router.py
@@ -6,7 +6,6 @@ from converters.schema_converters import (
 )
 from converters.schema_converters.schema_converter import SchemaConverter
 from database.model.dataset.dataset import Dataset
-
 from routers.resource_router import ResourceRouter
