+++ src/routers/resource_routers/educational_resource_router.py
