+++ src/routers/resource_routers/news_router.py
