+++ src/routers/resource_routers/platform_router.py
