+++ src/tests/routers/resource_routers/test_router_case_study.py
