+++ src/tests/routers/resource_routers/test_router_educational_resource.py
