+++ src/tests/routers/resource_routers/test_router_platform.py
